requirements.txt
tendo
pystray
keyboard
pillow
psutil
